00_basics-of-testing-with-jest/app/utils.ts
export class StringUtils {
  public toUpperCase(arg: string): string {
    if (!arg) {
      throw new Error('Invalid argument!');
    }
    return toUpperCase(arg);
  }
}

export const toUpperCase = (arg: string) => {
  return arg.toUpperCase();
};

export type StringInfo = {
  lowerCase: string;
  upperCase: string;
  characters: string[];
  length: number;
  extraInfo: Object | undefined;
};

export const getStringInfo = (arg: string): StringInfo => {
  return {
    lowerCase: arg.toLowerCase(),
    upperCase: arg.toUpperCase(),
    characters: Array.from(arg),
    length: arg.length,
    extraInfo: {},
  };
};
